+++ py_gnome/lib_gnome/TypeDefs.h
@@ -37,14 +37,12 @@ typedef short LandType;
 #ifdef _MSC_VER
 	#if _WIN64
 		// for now we will not change the type for windows64
-		//typedef long Seconds; // duration in seconds, or seconds since 1904 
-		typedef long long Seconds; // duration in seconds, or seconds since 1904  <--Bogdan Hlevca
+		typedef long Seconds; // duration in seconds, or seconds since 1904 
 	#else
 		#ifndef pyGNOME
 			typedef unsigned long Seconds;
 		#else
-			//typedef long Seconds; // duration in seconds, or seconds since 1904
-			typedef long long Seconds; // duration in seconds, or seconds since 1904  <--Bogdan Hlevca
+			typedef long Seconds; // duration in seconds, or seconds since 1904
 		#endif
 	#endif
 #else
